remove duplicate file
